EXECUTION_OUTPUT
As you can see, attendance before entry was a empty csv file. The live feed is basically the current feed from your webcam, or you can integrate it with another camera module. Attendance is captured as soon as a known student comes in the frame. Multiple faces can be recognized at once. Time of arrival is noted and attendance is marked.